requirements.txt
scrapy
tensorflow
streamlit
psycopg2-binary
pandas
numpy